# Improving Transaction Registry Performance