--- a/config/config.exs
+++ b/config/config.exs
@@ -1,6 +1,6 @@
 use Mix.Config
 
-if Mix.env() in [:test, :test_phoenix, :test_no_nif] do
+if Mix.env() in [:bench, :test, :test_phoenix, :test_no_nif] do
   config :logger,
     level: :warn,
     handle_otp_reports: false,