--- a/lib/appsignal/transaction/ets.ex
+++ b/lib/appsignal/transaction/ets.ex
@@ -9,9 +9,12 @@ defmodule Appsignal.Transaction.ETS do
   @table :"$appsignal_transaction_registry"
 
   def start_link do
-    Agent.start_link(fn ->
-      :ets.new(@table, [:set, :named_table, {:keypos, 1}, :public, {:write_concurrency, true}])
-    end, name: @name)
+    Agent.start_link(
+      fn ->
+        :ets.new(@table, [:set, :named_table, {:keypos, 1}, :public, {:write_concurrency, true}])
+      end,
+      name: @name
+    )
   end
 
   def insert(value), do: :ets.insert(@table, value)