This PR is all about improving the transaction registry. Currently, since it's a genserver, we end up getting backpressure. This PR is about trying to reduce that backpressure as intelligently as possible.

First up, we got a benchmark. I used the Fibonacci sequence to make a list of concurrent processes and then started hitting it. My initial benchmark for the system:

```
Name               ips        average  deviation         median         99th %
registry         0.109         9.15 s     ±0.00%         9.15 s         9.15 s
```

Then I checked in on what effect PR #503 had on the benchmark:

```
Name               ips        average  deviation         median         99th %
registry        0.0609        16.42 s     ±0.00%        16.42 s        16.42 s
```

Now, this is caused by silent backpressure, so to remove that we alter a `handle_cast` to a `handle_call` and we get:

```
Name               ips        average  deviation         median         99th %
registry         0.113         8.87 s     ±0.00%         8.87 s         8.87 s
```

Looking better, but barely. The problem is now with the GenServer in the transaction registry being the root cause of the backpressure.

The first step was moving ETS to its own file and its own process. We start the process up with an Agent to boot ETS and to make sure that ETS calls to its own process for management if needed in the future (such as changing ownership).

After that, the `init` call was free of responsibilities and we could start moving code over to a task; that's the `Receiver`. We start a simple task, register it, and listen and loop over messages send to the process. This will be the only blocking part.

Then we add two public functions called `monitor/1` and `demonitor/1` for the `Receiver` to monitor and demonitor. We update the code in the `TransactionRegistry` and benchmark:

```
Name               ips        average  deviation         median         99th %
registry         0.122         8.19 s     ±0.00%         8.19 s         8.19 s
```

The **average duration is 10.5% smaller** than the initial, the **ips is 10.6% larger**.