@BenMorganIO 2019-08-08
@tombruijn I would recommend relabelling this to critical performance fix ;)