@jeffkreeftmeijer 2019-08-09
Hey @BenMorganIO,

Thanks for taking the time to dive into this, and thanks for submitting such a well-polished pull request!

 We’re looking into this, but in order not to break any existing installs, we’ll need to do some thorough testing to make sure everything still functions as before to be able to merge this. We’ll keep you posted on our progress with this.

In the meantime, please feel free to run your fork, and please let us know if you run into any issues.