@BenMorganIO 2019-08-09
@ jeffkreeftmeijer FWIW, we haven't seen any new appsignal timeouts in our logs.