@jeffkreeftmeijer 2019-08-19
Hey @BenMorganIO, after merging some of your changes into develop already, I've opened a new pull request with the rest of your changes in #505 to get reviewed by the rest of the team. I'm closing this one, let's continue there.